Token TreeMaps
Re-implementation of a useful structue: the tree map. Given a list of
names and a token string, the tree map can find the longest name in the
list which starts at the current position of the token string. Very near
implementing actual parsing, as this was the last step in getting
GrammarTokens to work.